% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_duplicate_tax}
\alias{fix_duplicate_tax}
\title{fix_duplicate_tax}
\usage{
fix_duplicate_tax(physeq)
}
\arguments{
\item{physeq}{a phyloseq object}
}
\description{
fix the duplicatae taxonomy names of a phyloseq object
}
\author{
Chenghao Zhu, Chenhao Li, Guangchuang Yu
}
