% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatPhrase}
\alias{formatPhrase}
\title{formatPhrase}
\usage{
formatPhrase(sentence, taxon)
}
\arguments{
\item{taxon}{a taxon name to be italisized}

\item{phrase}{a phrase with taxon name(s)}
}
\description{
generate an expression for the phrase with the given taxon italisized
}
\author{
Chenhao Li, Guangchuang Yu
}
