% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{lefse.lda.plot}
\alias{lefse.lda.plot}
\title{lefse.lda.plot}
\usage{
lefse.lda.plot(
  lefse.result,
  negate.class = NULL,
  lda.threshold = NULL,
  group = NULL
)
}
\arguments{
\item{lefse.result}{the output file from lefse}

\item{negate.class}{the class to be shown on the negative side (default: everything show on the positive side)}

\item{lda.threshold}{features with LDA score less than this will be removed}

\item{group}{a data.frame with two columns for the subgroups of the classes (first column: the names of the features, second column: the subgroup these features belonging to)}
}
\description{
create a lefse LDA plot
}
\author{
Chenhao Li, Guangchuang Yu
}
