% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseMetaphlanTSV}
\alias{parseMetaphlanTSV}
\title{parseMetaphlanTSV}
\usage{
parseMetaphlanTSV(file, index = 1, delimiter = "\\\\|",
  node.size.scale = 1, node.size.offset = 1)
}
\arguments{
\item{file}{a single metaphlan table}

\item{index}{the column number of taxonomic annotation}

\item{delimiter}{delimiter to separate taxonomic anotations}

\item{node.size.scale}{the parameter 'a' controlling node size: nodeSize=a*log(relative_abundance)+b}

\item{node.size.offset}{the parameter 'b' controlling node size: nodeSize=a*log(relative_abundance)+b}
}
\value{
a treeio::treedata object
}
\description{
parse a MetaPhlan table to a tree object
}
\author{
Chenhao Li, Guangchuang Yu
}
