% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS014459_Stool_profile.R
\docType{data}
\name{SRS014459_Stool_profile}
\alias{SRS014459_Stool_profile}
\title{A sample metaphlan table}
\format{A data frame with 33 rows and 2 variables:
\describe{
  \item{V1}{Taxonomic annotation}
  \item{V2}{Relative abundances}
  ...
}}
\source{
\url{https://bitbucket.org/biobakery/biobakery/raw/tip/demos/biobakery_demos/data/metaphlan2/output/SRS014459-Stool_profile.txt}
}
\usage{
SRS014459_Stool_profile
}
\description{
A sample metaphlan table
}
\keyword{datasets}
