% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_taxa}
\alias{summarize_taxa}
\title{summarize_taxa}
\usage{
summarize_taxa(physeq, level, keep_full_tax = TRUE)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{level}{the taxonomy level to summarize}
}
\description{
Summarize a phyloseq object on different taxonomy level
}
\author{
Chenghao Zhu, Chenhao Li, Guangchuang Yu
}
