% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{tree.backbone}
\alias{tree.backbone}
\title{tree.backbone}
\usage{
tree.backbone(tree, size = 2, layout = "circular", shape = 21,
  fill = "white")
}
\arguments{
\item{tree}{a treeio::treedata object}

\item{size}{branch width}

\item{layout}{tree layout}

\item{shape}{clade node shape}

\item{fill}{clade node fill}
}
\value{
a ggtree object
}
\description{
basic tree (backbone) plotting utility
}
\author{
Chenhao Li, Guangchuang Yu
}
