% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{clade.anno}
\alias{clade.anno}
\title{clade.anno}
\usage{
clade.anno(gtree, anno.data)
}
\arguments{
\item{gtree}{a ggtree object}

\item{anno.data}{a 2 column data.frame of annotation information. It has columns of clade name and color used for highlighting.}
}
\value{
a ggtree object
}
\description{
annotate a ggtree plot to highlight certain clades
}
\author{
Chenhao Li, Guangchuang Yu
}
