% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseMetaphlanTSV}
\alias{parseMetaphlanTSV}
\title{parseMetaphlanTSV}
\usage{
parseMetaphlanTSV(
  tax.profile,
  index = 1,
  header = FALSE,
  delimiter = "\\\\|",
  node.size.scale = 1,
  node.size.offset = 1
)
}
\arguments{
\item{tax.profile}{a single metaphlan table (either a file path or a a data.frame)}

\item{index}{the column number of taxonomic annotation}

\item{header}{whether tax.profile provided as a file path contains a header}

\item{delimiter}{delimiter to separate taxonomic anotations}

\item{node.size.scale}{the parameter 'a' controlling node size: nodeSize=a*log(relative_abundance)+b}

\item{node.size.offset}{the parameter 'b' controlling node size: nodeSize=a*log(relative_abundance)+b}
}
\value{
a treeio::treedata object
}
\description{
parse a MetaPhlan table to a tree object
}
\author{
Chenhao Li, Guangchuang Yu
}
