% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseMetaphlanTSV}
\alias{parseMetaphlanTSV}
\title{parseMetaphlanTSV}
\usage{
parseMetaphlanTSV(file, index = 1, delimiter = "\\\\|")
}
\arguments{
\item{file}{a single metaphlan table}

\item{index}{the column number of taxonomic annotation}

\item{delimiter}{delimiter to separate taxonomic anotations}
}
\value{
a phylo4d object
}
\description{
parse a metaphlan table to a tree object
}
\author{
Chenhao Li
}
